requirements.txt
autopep8==2.0.4
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
Pillow==10.1.0
pycodestyle==2.11.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
six==1.16.0
tomli==2.0.1
urllib3==2.1.0
